move files to template directory, remove cmake for Python
