+++ {{cookiecutter.repo_name}}/templates/cpp/CMakeLists.txt
