+++ {{cookiecutter.repo_name}}/templates/cpp/STUBS_MPI/mpi.h
