+++ {{cookiecutter.repo_name}}/templates/mdimechanic_python.yml
@@ -1,0 +1,19 @@
+code_name: '{{ cookiecutter.repo_name }}'
+docker:
+  image_name: '{{ cookiecutter.repo_name }}'
+
+  build_image:
+    - ""
+
+  build_code:
+    - pip install -r {{ cookiecutter.repo_name }}/requirements.txt
+
+  validate_engine:
+    - echo "Insert code that will confirm that your code has been built successfully"
+    - exit 1
+
+engine_tests:
+  # Provide at least one example input that can be used to test your code's MDI functionality
+  script:
+    - echo "Insert commands to run an example calculation here"
+    - exit 1
