+++ {{cookiecutter.repo_name}}/templates/python/requirements.txt
@@ -1,0 +1,1 @@
+pymdi
